refactor: poem Preprocessing
